Move types out of MVC
