+++ src/Http/Http.Abstractions/src/Metadata/ITagsMetadata.cs
@@ -1,7 +1,7 @@
 // Licensed to the .NET Foundation under one or more agreements.
 // The .NET Foundation licenses this file to you under the MIT license.
 
-namespace Microsoft.AspNetCore.Http
+namespace Microsoft.AspNetCore.Http.Metadata
 {
     /// <summary>
     /// Defines a contract used to specify a collection of tags in <see cref="Endpoint.Metadata"/>.
